Had planned to make the progrom prepend 16 prime numbers to the beginning of the file to be encrypted, and they'd be check for primality after decryption to verify that the process had been successful.  I realized, however, that this was flawed in that one could easily, having seen the source code, crack the code.  How?
For each of the first 16 bytes, run an xor operation using a prime number,
then run the output against the original byte, and if the final output is
the prime number you started with, the output from the first xor operation
is a character in the encryption key.

So, in other words, back to the drawing board.  I suspect that the most
secure way to verify the decryption key is to prepend a bcrypt hash to the
beginning of the file, and simply authenicate the password using said hash.
Obviously, this bcrypt hash should only be used for password verification,
not the actual encryption process.  A separate hash should be used for this.